* origin/develop: (44 commits)
  Removed SDLTTSChunkFactory from project.
  Removed SDLRPCRequestFactory from project.
  Remove a test focus
  Fixed test cases
  Fix podspec not updated with latest version number
  Update for v4.5.0
  Follow style guide for boolean check
  Updated formatting of README.md
  Removed nonexistent implementation file from project.
  Added steps for using the lock screen manager from SDL's bundle to the readme.
  Added check when trying to upload an appIcon for a head unit that does not support graphics.
  Update for v4.5.0
  Fixed issue with trying to upload a file that is nil will cause an infinite loop.
  Fixed an issue with initializer causing infinite loop.
  Improved readability for nil-checks.
  Fixed issue relating to SDLLocationDetails' searchAddress using the wrong key for object store.
  Fixed formatting for some interfaces.
  Fixed issue with LocationDetails' properties not properly becoming initialized.
  Fixed issue with waypoints not being correctly initialized as SDLLocationDetails
  Adding in nil checks for completion handlers of Streaming Media Manager.
  ...

# Conflicts:
#	SmartDeviceLink/SDLDateTime.m
#	SmartDeviceLink/SDLGetWaypoints.m
#	SmartDeviceLink/SDLGetWaypointsResponse.m
#	SmartDeviceLink/SDLLocationCoordinate.m
#	SmartDeviceLink/SDLLocationDetails.m
#	SmartDeviceLink/SDLOasisAddress.m
#	SmartDeviceLink/SDLOnWaypointChange.m
#	SmartDeviceLink/SDLSendLocation.m
#	SmartDeviceLink/SDLUnsubscribeWaypoints.m
#	SmartDeviceLink/SDLUnsubscribeWaypointsResponse.m
#	SmartDeviceLinkTests/RPCSpecs/StructSpecs/SDLLocationDetailsSpec.m

+++ SmartDeviceLink/SDLAddSubMenu.m
@@ -13,7 +13,7 @@
 }
 
 
-- (instancetype)initWithId:(UInt8)menuId menuName:(NSString *)menuName position:(UInt8)position {
+- (instancetype)initWithId:(UInt32)menuId menuName:(NSString *)menuName position:(UInt8)position {
     self = [self initWithId:menuId menuName:menuName];
     if (!self) {
         return nil;
@@ -24,7 +24,7 @@
     return self;
 }
 
-- (instancetype)initWithId:(UInt8)menuId menuName:(NSString *)menuName {
+- (instancetype)initWithId:(UInt32)menuId menuName:(NSString *)menuName {
     self = [self init];
     if (!self) {
         return nil;
