+++ SmartDeviceLink/SDLSetMediaClockTimer.h
@@ -19,7 +19,7 @@
  */
 @interface SDLSetMediaClockTimer : SDLRPCRequest
 
-- (instancetype)initWithUpdateMode:(SDLUpdateMode)updateMode hours:(NSInteger)hours minutes:(NSInteger)minutes seconds:(NSInteger)seconds;
+- (instancetype)initWithUpdateMode:(SDLUpdateMode)updateMode hours:(UInt8)hours minutes:(UInt8)minutes seconds:(UInt8)seconds;
 
 - (instancetype)initWithUpdateMode:(SDLUpdateMode)updateMode;
